HTCondor: do not include 'Environment = ' line by default
If no 'env_extra' argument is supplied by the user, the line
'Environment = ""' will still be included in the HTCondor submission file.
This overrides the option 'getenv = True', which can be used to
propagate the full current environment of the user to the worker jobs.

With this commit, the line 'Environment = ""' is not added if 'env_extra'
is not given. Keeping this line can be done by specifying 'env_extra=[]'.

By default, HTCondor does not propagate the current environment to the
jobs, so that the current default behaviour is not modified.